backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-auth==2.35.0
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.40.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
